maintt/requirements.txt
click==6.6
Django==1.10.2
djangorestframework==3.4.7
fake-factory==0.7.2
ipaddress==1.0.17
mysqlclient==1.3.9
python-dateutil==2.5.3
python-dotenv==0.6.0
six==1.10.0
wheel==0.24.0
